requirements.txt
python-telegram-bot
sentence-transformers
faiss-cpu
pandas